--- a/Cargo.toml
+++ b/Cargo.toml
@@ -11,6 +11,7 @@ description = "A hash table with consistent order and fast iteration."
 bench = false
 
 [dependencies]
+efficient_enum = { version = "^0.3.1", optional = true }
 
 [dev-dependencies]
 itertools = "0.5.1"
@@ -23,6 +24,10 @@ lazy_static = "0.2"
 # for testing only, of course
 test_low_transition_point = []
 test_debug = []
+test_efficient_enum = ["efficient_enum", "unsafe"]
+test_unsafe_efficient_enum = ["test_efficient_enum"]
+test_unsafe_index = ["unsafe"]
+unsafe = []
 
 [profile.bench]
 debug = true