--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,9 +10,6 @@ description = "A hash table with consistent order and fast iteration."
 [lib]
 bench = false
 
-[dependencies]
-efficient_enum = { version = "^0.3.1", optional = true }
-
 [dev-dependencies]
 itertools = "0.5.1"
 rand = "0.3"
@@ -24,10 +21,6 @@ lazy_static = "0.2"
 # for testing only, of course
 test_low_transition_point = []
 test_debug = []
-test_efficient_enum = ["efficient_enum", "unsafe"]
-test_unsafe_efficient_enum = ["test_efficient_enum"]
-test_unsafe_index = ["unsafe"]
-unsafe = []
 
 [profile.bench]
 debug = true