.retain() now returns an iterator with the removed items
closes bluss/ordermap#11
Also cleaned up some match statements.