@bluss 2016-11-15
Thanks, that's interesting. It seems like a direct translation of the retain method into an iterator.

You're kind of putting me on the spot here, I have to review the design of the retain iterator and if I do it with my usual critical eye, there are some imperfections here:

- OrderMap gets an iterator that does not use the regular order of the elements
- Run on Drop is a questionable design pattern in itself. (Open question: Why should we use this pattern if we are making a new API?)
  - `forget` problematics show how nonintuitive it can be to tie behaviour to a temporary value
  - the operation is lazy as an iterator but eager on drop.
- I believe this crate can implement an in-order "retain" iterator with the same algorithmic complexity