@Popog 2016-11-15
Well, that was why I said it seemed simple 😝 .

The only alternative I can think of to Run on Drop is how `thread::scoped` solved its issues. I'm certainly open to:
```rust
pub fn retain<F, G, R>(&mut self, keep: F, retain: G) -> R
    where F: FnMut(&mut K, &mut V) -> bool,
    G: FnOnce(Retain<K, V, S, F>) -> R
```

Which would complete the operation before returning.  I suspect `forget` would prevent, or at least significantly slow in-order traversal. So this would avoid that, and still allow users to do simple loops or `.collect()` it for storage.