@Popog 2016-11-15
That's somewhat unintuitive for users who don't want to iterate (i.e. people who would like to use the current function):
```rust
m.retain(|k, _| k < 5); // does nothing
m.retain(|k, _| k < 5).count(); // actually works
```