@bluss 2016-11-15
Some ideas, just in that case:

1. Rename the method so that expectations change (can't keep the "retain" name since the expectations from existing APIs is wrong).
2. use `#[must_use]` so that it makes a warning when it does nothing (just like iterator adaptors like .map())
3. add an explicit method for draining the iterator / running it til the end.