@Popog 2016-11-15
I think I'm now thoroughly convinced you cannot have all of these:
* O(n)
* A forgettable return value (no Drop or scoped style impl).
* No change to iteration order.

There maybe some exceedingly clever way of accomplishing this, but if so, I'm not seeing it.