@Popog 2016-11-17
> 1. Rename the method so that expectations change (can't keep the "retain" name since the expectations from existing APIs is wrong).

Name idea `filter_cursor`