@Popog 2016-11-25
Well, this no longer adds a `Retain` iterator, but it does add tombstones, which will make implementing `filter_cursor` trivially easy.