@Popog 2016-11-25
I've only changed `Entry` and `retain` to use tombstones, `swap_remove` and any benches which test that are unaffected. Overall, very little change. Tested on an i7 4770k:

name | original ns/iter | tombstone ns/iter
--- | --- | ---
entry_hashmap_150 | 3,209 (± 408) | 3,097 (± 60)
entry_orderedmap_150 | 2,846 (± 56) | 3,693 (± 48)
grow_fnv_hashmap_100_000 | 3,822,576 (± 134,932) | 3,862,308 (± 377,264)
grow_fnv_ordermap_100_000 | 4,027,347 (± 264,053) | 5,034,100 (± 735,180)
hashmap_merge_shuffle | 1,035,783 (± 49,163) | 1,064,027 (± 28,407)
hashmap_merge_simple | 34,034,802 (± 445,112) | 33,043,395 (± 355,684)
insert_hashmap_100_000 | 4,467,652 (± 1,126,408) | 4,277,086 (± 424,272)
insert_hashmap_10_000 | 267,512 (± 10,203) | 264,952 (± 7,215)
insert_hashmap_150 | 3,330 (± 63) | 3,220 (± 37)
insert_hashmap_int_bigvalue_10_000 | 644,812 (± 58,041) | 646,611 (± 13,821)
insert_hashmap_str_10_000 | 285,223 (± 5,886) | 284,730 (± 6,332)
insert_hashmap_string_10_000 | 1,229,139 (± 43,086) | 1,214,361 (± 22,182)
insert_orderedmap_100_000 | 3,656,966 (± 2,233,920) | 4,410,542 (± 811,701)
insert_orderedmap_10_000 | 251,699 (± 52,062) | 292,366 (± 15,493)
insert_orderedmap_150 | 3,269 (± 87) | 3,831 (± 19)
insert_orderedmap_int_bigvalue_10_000 | 458,839 (± 16,179) | 494,504 (± 13,260)
insert_orderedmap_str_10_000 | 265,618 (± 5,100) | 288,197 (± 2,769)
insert_orderedmap_string_10_000 | 1,083,413 (± 8,313) | 1,112,794 (± 14,495)
iter_black_box_hashmap_10_000 | 41,517 (± 572) | 44,543 (± 773)
iter_black_box_orderedmap_10_000 | 2,965 (± 24) | 5,509 (± 45)
iter_sum_hashmap_10_000 | 46,132 (± 1,379) | 44,069 (± 192)
iter_sum_orderedmap_10_000 | 2,594 (± 68) | 5,465 (± 101)
lookup_hashmap_100_000_inorder_multi | 164,799 (± 1,983) | 164,095 (± 1,107)
lookup_hashmap_100_000_multi | 170,325 (± 2,209) | 171,914 (± 5,477)
lookup_hashmap_100_000_single | 36 (± 8) | 35 (± 1)
lookup_hashmap_10_000_exist | 130,540 (± 8,240) | 127,333 (± 4,324)
lookup_hashmap_10_000_noexist | 135,121 (± 1,772) | 130,829 (± 1,057)
lookup_orderedmap_10_000_exist | 135,564 (± 2,158) | 145,417 (± 1,710)
lookup_orderedmap_10_000_noexist | 130,561 (± 2,740) | 136,118 (± 3,032)
lookup_ordermap_100_000_inorder_multi | 110,872 (± 1,311) | 117,463 (± 2,315)
lookup_ordermap_100_000_multi | 158,997 (± 3,455) | 159,648 (± 1,922)
lookup_ordermap_100_000_single | 33 (± 33) | 33 (± 10)
new_hashmap | 7 (± 0) | 7 (± 0)
new_orderedmap | 11 (± 0) | 14 (± 0)
ordermap_merge_shuffle | 646,260 (± 16,298) | 977,019 (± 5,665)
ordermap_merge_simple | 430,928 (± 36,904) | 786,512 (± 7,431)
pop_ordermap_100_000 | 2,989,304 (± 151,554) | 2,588,539 (± 49,431)
remove_ordermap_100_000 | 6,585,708 (± 447,388) | 5,901,486 (± 1,301,382)
retain_few_ordermap_100_000 | 3,533,252 (± 132,025) | 2,266,023 (± 67,881)
retain_half_ordermap_100_000 | 3,127,368 (± 608,248) | 2,035,521 (± 39,373)
retain_many_ordermap_100_000 | 1,127,635 (± 115,684) | 1,299,009 (± 22,572)
with_capacity_10e5_hashmap | 5,150 (± 148) | 5,115 (± 19)
with_capacity_10e5_orderedmap | 5,454 (± 130) | 5,485 (± 47)
insert_hashmap_string_10_000 | 1,257,354 (± 60,714) | 1,229,721 (± 7,204)
insert_hashmap_string_oneshot_10_000 | 1,173,116 (± 67,376) | 1,149,863 (± 5,214)
insert_orderedmap_string_10_000 | 1,093,339 (± 23,124) | 1,092,168 (± 13,019)
lookup_hashmap_10_000_exist_string | 190,106 (± 2,865) | 190,600 (± 2,763)
lookup_hashmap_10_000_exist_string_oneshot | 171,826 (± 54,066) | 170,334 (± 2,971)
lookup_ordermap_10_000_exist_string | 217,790 (± 4,591) | 224,351 (± 1,487)
lookup_ordermap_10_000_exist_string_oneshot | 197,931 (± 7,179) | 192,204 (± 2,247)

I'm going to investigate using the MSB of the hash in place of `Option (requires `unsafe` [either directly or as a dependency]) as I have a hunch that will probably provide some perf benefits.