@Popog 2016-12-02
Found some bugs and fixed them. Also added some unsafe options under test feature flags.  They don't really seem to provide any perf benefit, so I'll probably end up scrapping them.

The CI break seems to be from an SSL error for crates.io trying to pull the tagged union packaged I added as a dependency. Not sure what's up with that but with how things are looking I'll probably end up removing the dependency.