@Popog 2016-12-05
So, found a couple more bugs, but while I'm already touching so much of the code, how would you feel about the following changes:
* Remove `self.mask` in favor of getting the mask via `indices.len() - 1`
* Removing `probe_loop!` in favor of `loop` and `probe = (probe + 1) & mask`

My justification for the former is, it seems like a bad idea to have duplicate data that could get out of sync. The mask is already implicitly stored in `indices` and is cheap to recover.

My justification for the later is it reads better than the macro (especially when debugging as line numbers get collapsed to the first line of the macro) and seems to have very little perf cost.