@bluss 2016-12-06
I'm sorry I haven't sat down and looked at this yet. I would hold off on both of those changes.